[refactor] Run multiloop test inside Pytester to avoid the custom event loop implementation to pollute the test environment.
